Moved repository listing to separate component
